Refactor and put things in places
